Allow Infidel quest leader to demand the quest artifact.
I've thought about this for some time now, I don't think it's fair that
the Infidel role should get a free pass because their quest artifact is
needed to win the game. It's not like it was destroyed, it just has the
potential to change hands. And being the evil, sinister Infidel that you
are, you're most likely thinking of a way to kill your quest leader
anyways. In lieu of this, I've adjusted the alignment penalty an Infidel
gets for killing their pet, it's still a knock on their alignment, but
not as bad as that for other roles (this should have been a thing
already).