Fix: bug allowing potential bypass of most of the elemental planes.
From GruntHack - when I pulled over randomized elemental planes from
GruntHack, I neglected to also include a change that would take the
player to the first planes level (whatever that might be for that game)
instead of going directly to the plane of earth when quaffing a cursed
potion of gain level while holding the Amulet of Yendor. This bug has
been present for quite a while.