Alternate ending to the Ice Queen branch if the player is an Infidel.
Kathryn the Enchantress was placed under her curse by a powerful, evil
witch in the service of Moloch. So it stands to reason that she should
not be happy about having an Infidel free her from that curse. So here's
what happens at the end of the ice queen branch if the player is an
Infidel - Kathryn stays hostile, and the message she gives to the player
after being freed has changed. The player then has 75 turns to get the
hell out of Dodge before the Enchantress ends her grace period and
starts to attack. 75 turns is ample time to free the captive pegasus and
loot enough of the chests to get all the high-end gear, and still make
it out in time.