Ensure genocide applies to racial monsters
Currently I don't think this is possible to actually achieve, because
all the "racial templates" being used cannot be genocided.  But it's
theoretically possible to apply any monster type as a "race" to another
(e.g. a newt soldier), and in that case you'd want to ensure the newt
soldier gets killed when the player genocides newts.  This may be
unnecessary futureproofing -- what are the chances of newt soldiers,
after all -- but better safe than sorry.