Revert ball & chain being rustproof, prevent either from rusting away completely.
Eariler steps taken to prevent the multitude of crashes that could occur
if a heavy iron ball and/or its chain were to rust away completely did
not quite work out. This commit does two things: allows both the ball
and chain to be able to rust, but then does not allow them to rust
completely away like other unprotected iron objects. This is a
short-term preventative measure to eliminate any future crashes until a
better solution is found that makes the ball & chain fair game to being
able to rust completely away without crashing the game.