Fix: pet unable to "return" item to shop.
Currently a PR sitting in the xNetHack queue.

'Picking up an item in a shop, using #loot to give it to your pet, then
allowing your pet to drop the item, would cause the item to remain on
your bill despite it having been returned to the shop inventory. If a
pet drops an unpaid item in a shop, ensure it is removed from the bill.'

Co-authored-by: entrez <me@entrez.cc>