Adjust how often Grimtooth can disease its target.
Hate to say it, but this is an actual nerf. I came to the realization
that Grimtooth would cause disease/terminal illness to its target on
every single hit. Poisoned weapons don't come anywhere near that, it's a
1 in 10 chance for those to instakill their target. Because very few
monsters are sick resistant, being infected by Grimtooth is a virtual
death sentence. I've reduced the odds of infection from 100% down to
roughly 17%. Grimtooth is still a very deadly weapon, but as of now I
don't feel like it's overpowered.