Fix: super tame steeds attacking monsters they shouldn't while being ridden.
A while ago I set a condition to where if your steed had a tameness
value of 15 or higher, it would attack hostiles on your behalf instead
of waiting to be attacked first. Unfortunately this bypassed
acceptable_pet_target(), meaning your awesome super tame warhorse would
decide to kick that cockatrice for the lulz. This commit addresses that.
FYI, if it's directly attacked, your steed will fight back.