Fix: prevent select_newcham_form() from choosing player monsters as a valid polyform (except for doppelgangers).
Because of player monsters' placement in monst.c before 'long worm
tail', it allows them to spawn randomly, but it allowed them to be a
valid choice for select_newcham_form(), say from a monster or pet
walking over a poly trap. This commit addresses that, as that shouldn't
happen, exception being doppelgangers. Thanks qt.

Co-authored-by: entrez <me@entrez.cc>