Change dlord() to ndemon() for sacrificing same race on chaotic or unaligned altar.
When I added more demon lords and princes to EvilHack, I also redid what
was considered a demon lord vs a demon prince (devteam had this all
wrong according to dungeons & dragons lore). because of that, there are
now many more demon lords and a lot fewer demon princes. Meaning if a
chaotic or unaligned (infidel) player were to perform a same race
sacrifice on a co-aligned altar, they'd be able to summon virtually
every demon lord in the game. This was definitely not the intent, there
needs to be a good chance that the player will encounter a demon lord in
the first two demon lairs in Gehennom. Lesser demons have a chance to
gate in a demon lord of course, but those are low odds and it's not
something the player has any control over - I'm fine with that.

This commit swaps out dlord() with ndemon() when a same race sacrifice
is performed on a co-aligned altar for chaotic/unaligned players. In
vanilla, the demon summoned is always peaceful - here there's a 20%
chance the demon is hostile, as demons typically aren't fond of being
summoned against their will.