Adjustments to mspec_used for boss-level monsters, Wizard of Yendor gear tweaks.
Based on player feedback now that version 0.7.0 has seen a decent amount
of playtime and one ascension. The Wizard of Yendor's buffs may have
been a bit overboard. The biggest impact was mspec_used being 0 all of
the time for certain boss-type monsters, including the Wizard. That
coupled with him being able to cast clerical-based spells on top of
wizard-based spells was a bit much. Wouldn't be that big of a deal, but
since the Wizard can cast double-trouble and reappears throughout the
ascension run, some adjustments needed to be made.

* The Wizard of Yendor having a chance for a weapon with the shock
object property is changed to venom object property
* Since the Wizard already has a decent base AC, adjust the enchantment
of the armor he spawns with
* mpsec_used 0 before was 2/3rds of the time for demon lords, quest
leaders/nemesis, high priests and Kathryn the Ice Queen, and all of time
for demon princes, the Wizard, Vecna, and Kathryn the Enchantress. It's
now changed to half of the time for all listed across the board

There's also a comment fix in here for the Eye of Vecna.