Adjust how often Vecna warps to the upstairs.
Any monster that covets either the Amulet of Yendor or any of the
invocation items will warp to the upstairs if their hit points reach a
certain level. A couple players that have encountered Vecna got the
impression that this was happening too soon. If you look at strategy()
in wizard.c, this can occur more often than you might expect. There's a
bit in here specifically for the Wizard of Yendor to prevent him from
doing that as often as other monsters. This commit adds Vecna to that
bit as well - flavor-wise this makes sense. Both are uber-powerful
wizards with an inflated ego that think they are indestructable.