Readjust alignment penalties for Knights (again).
Way back when EvilHack was new, I had incorporated SporkHack's 'Knight
patch'. One of the changes in that patch was to drastically change
alignment penalties for certain things, like eating while satiated, or
attacking with a poisoned weapon. In light of the new alignment abuse
tracking for the quest, I've toned those values down a bit so it's not
so crazy. There's also a fair amount of code formatting cleanup here
also.