Fix: zombie form into a zombie that has been genocided.
If a type of zombie had been genocided, say orc zombies, and then later
on an orc is killed by a zombie, they'd try to arise again as an orc
zombie. Except they couldn't because orc zombies don't exist anymore,
and it just went downhill from there. This commit addresses this by
checking to see if the zombie form the affected monster wants to turn
into has already been genocided - if so, just die outright instead of
dying then arising again as a zombie.

Some other edits here as well to check race instead of the base monster
(thanks qt).