Fix: crowned Infidel giants could not wear body armor/cloaks.
An Infidel giant, once crowned, turns in a demon. This demon is
human-sized, and they would lose all of the abilities of being a giant
(good and bad) save one, that being unable to wear body armor or cloaks.
That's fixed in this commit.

Also corrected some minor bits concerning hands vs claws (demons and
illithids both have claws, so reflect that when they attack or
wield/wear objects).