Fix: crash due to incorrect handling of rot timer.
This was self-induced from a recent commit that prevented corpses looted
out of an ice box by monsters from not rotting. However, a couple of
defines were reversed from a PR a long time ago (git commit 8201466)
that broke this recent fix. This commit addresses that.

One quirk that I can live with for now - if a food item becomes burnt,
it will show as 'rotten'.