Fix: prevent shallow water in mindflayer nurseries from forming on stairs.
Main reason being, if you dip something into shallow water, there's a
chance the puddle could disappear, and if said puddle is on stairs, it
could make the stairs go away (this will be addressed in the next
commit). This fix also has better handling of determining what makes
this room a 'special nursery' (the neothelid rampage).

Co-authored-by: entrez <me@entrez.cc>