Adjust maximum hit points the Wand of Orcus can decrease per hit.
The initial implementation was a bit over the top, didn't realize by how
much until actually experiencing it during a real game. The tweaks in
this commit takes the edge off a bit. Orcus should still be taken
seriously, and getting hit with the Wand of Orcus is still considered
a thing you don't want to have happen. But now it should just make
players exclaim 'OMG' instead of 'OMFG'.