Fix: getting incorrect damage feedback from poisoned/diseased weapons if fully resistant to their special attack.
This was initially reported in regards to being hit by the Sword of Kas,
but in looking at this, I noticed this behavior was also occuring with
weapons with the ITEM_VENOM object property, as well as Grimtooth. If
the player is fully resistant to the special attack, the feedback should
reflect that.