Fix: monster reading a scroll of remove curse.
Basically, this wasn't working correctly at all for worn/wielded objects
in a monsters inventory. Also added a condition, if the scroll is
cursed, don't attempt to use it as a cursed scroll of remove curse
doesn't do anything except unpunish (ball & chain) and monsters are not
subject to that.