Generate monsters on branch stairs after performing the invocation.
From UnNetHack git commit 21f1f57 (modified):

'This was an obvious oversight of the monster generation code taken from
SporkHack.

While the SporkHack code generated most monsters on the upstairs, it
didn't do this for branch upstairs. Although for a normal game this
should only affect dungeon level 1 and the VotD.'