Infidel role and object properties/materials
Something I would've added in my own patch had vanilla NetHack had these
features.  First, the Paladin's weapon has a chance to be silver, and a
further chance to receive a frost or shock property.  She isn't quite
important enough to have an artifact, but this is a good middle ground.

Second, as Moloch also doesn't have a crowning artifact, he instead may
add the fire property to a wielded (non-artifact) weapon (only if the
player already knows the gifted spell, as usual).  This is slightly
controversial, as it can create a weapon with two elemental properties
at once; but at least it's only a maybe-once-per-game opportunity.