Fix: zapping at an object with teleportation resulted in scrambled zap targetting as soon as any object was hit.
From NetHack 3.7 git commit 475648b:

'flooreffects() got changed to set bhitpos for handling erosion damage,
but bhitpos is used by wand/spell zap handling to pick the next target
spot.  So a zap that teleported an object resumed one step beyond the
object's destination rather than one step beyond where it was hit.
The rest of the zap could hit things (monsters as well as other
objects) which weren't in line with the original zap.  Reported case
was one of the Riders on Astral, coming from beyond multiple walls to
arrive adjacent to hero (usual position when a Rider gets teleported).'

Because EvilHack has some of the bhitpos fixes from NetHack 3.7, this
bug affected EvilHack as well. Never had a report of one of the Riders
appearing next to the player on Astral as described in the original
commit, but it could have happened.