Monsters can use potions of oil as an offensive item.
From xNetHack git commit 9a55ff8:

'The muse code for oil follows most other potions, with the added
wrinkle that in order to make it explode, it needs to get lit before the
monster throws it (then the rest of the stack, if any, needs to get
snuffed afterward). There are also some rules in place to prevent them
from selecting oil to throw if it's likely to fatally blow themselves
up, too. (Though they don't check to see if they're also blowing up
their own allies.)

Related features in this commit include:
- if non-burning oil splatters over you, it of course doesn't explode,
  but it makes your hands slippery.
- a small chance for monsters above difficulty level 6 to receive oil as
  a defensive item (stealing 1/30 of the probability from wands of fire)
- additional cases in which oil auto-identifies (if you see a spot where
  oil explodes, or an unlit dknown potion splatters oil on you)
- monsters that can use items will take oil if they find it'

When I ascended xNetHack last month, I was able to experience this
behavior firsthand, thought it was pretty neat. Nice addition to the
monster AI.