Fix: various issues with monster riders/steeds.
This commit addresses a couple issues with monster riders/steeds. It
fixes odd behavior seen when having a monster rider/steed present,
saving the game, reloading and seeing the rider/steed separated. It also
addresses a condition where a monster rider could be on two steeds at
once, causing one of the steeds to look as if it had an invisible or
non-existent rider. Thanks qt.

Co-authored-by: entrez <me@entrez.cc>