Fix: give rings/wands/artifacts in monsters inventory the same consideration as the player vs shock damage.
Git commit 18d071a afforded wands/rings in the players inventory a
chance to resist shock damage (artifacts were already off-limits via
destroy_item()), but none of that was carried over to the same objects
in a monsters inventory. Wands and rings now have the same chances to
resist based on their material/fixed status in a monsters inventory.
Artifacts in a monsters inventory are off-limits, same as for players.
Kinda surprised that bit isn't in vanilla NetHack, but the odds of
attacking a monster with a lightning-based attack are much less there.