Adjust monster spawns in morgues in Vecna's domain.
It's been pointed out that because of the sheer size of the morgues on
Vecna's first level, the number of wraith spawns was so great that it
was basically a guarantee to fast-track to experience level 30. I'm not
against farming wraith corpses to gain experience, but it should not be
this easy to do. This commit tones down the chances for a wraith to
spawn in a morgue in Vecna's domain. It also allows for shades to spawn
here as well. This does NOT affect the Valley of the Dead in any way.