Fix: include steed's worn barding when tossing potions at it.
Discovered an oversight when implementing barding - there aren't many
ways to deal with a cursed saddle being worn. One of the few ways of
dealing with it, is to throw potions of holy water at your steed and
hope that it lands a hit on its saddle, uncursing it. I neglected to add
the same for worn barding. This commit addresses that omission.