Fix: prevent thievery skill from incrementing when only thing in the monster's inventory is gold.
A couple years ago, the devteam made it so the steal_it() function did
not steal gold from a monsters inventory (see git commit 58583ca). It's
a bit funky how that works - the gold is seen ininventory, it's then
taken out before the actual steal objects routine is ran, then it puts
the gold back into the monsters inventory. The thievery function would
see that there was still something (in this case, just gold pieces) in a
monster's inventory before steal_it() was called - even though nothing
was stolen, the skill would be incremented/trained. This commit prevents
that from occurring.

I'm not happy with this, as the player will still receive 'You attempt
to steal from <foo>' feedback, but nothing afterwards. Make you think
you should have stolen something, but didn't. Not sure whether to
somehow work in feedback for gold specifically (except it's not in their
inventory for that brief slice of time), or revert the devteam commit
and make gold fair game.