Ability to choose worn piece of armor to enchant/repair.
This commit gives the player the ability to choose which piece of worn
armor to enchant (scroll of enchant armor) or repair (spell 'repair
armor'). Some caveats:

* armor must be worn to be able to specify a piece to enchant/repair
* reading a scroll of enchant armor and not knowing what type of scroll
it was initially will exhibit past behavior, enchanting a worn piece of
armor at random
* Casting the spell 'repair armor', the caster must have basic skill or
greater in 'matter' school of spells to be able to choose a worn piece
of armor to repair, otherwise one is chosen at random

Thanks amateurhour for the suggestion and initial patch.

Co-authored-by: gebulmer <gebulmer@googlemail.com>