Fix: air terrain in Vlad's cavern and is_clinger().
Follow up to the last commit. Monsters that are part of is_clinger()
were not taken into account - they should be able to cling to the
ceiling in Vlad's cavern and move across it. One thing I haven't worked
out yet is being able to successfully ride a monster that can cling
(wumpus) in the open air in Vlad's cavern without falling off. Another
fix in this commit, proper handling of your steed should you dismount it
in open air (Vlad's) and the only way it was not falling was because of
the players status (flying/lev).