Fix: barrow wight's spell attack should be last in the chain.
Players that polymorph into monsters that can cast spells, can use those
spells as well as they attack (in vanilla NetHack, the spell attack is
swapped out for another type of physical attack). If a player had poly'd
into a barrow wight, the last attack type in the chain was a claw
attack, which would still fire if the monster they were attacking was
killed by the previous spell attack in the same round, causing a
dmonsfree error. Swap the attack order so spells are last in the chain.

There is probably a better way to do this, since physical attacks will
halt once the player's target is dead. Vanilla NetHack got around this
by just changing any monster spell attacks to some other physical type
instead of allowing the player to actually use monster spells. Something
to look further into for sure.