Fix: polymorphed player (by monster) and their inventory.
With the inclusion of monsters being able to use wands of polymorph
against the player, an issue was discovered that when the player is
zapped and they polymorph, dropped objects in their inventory were also
subject to being polymorphed in the same turn. The assumption was that
objects in inventory were safe (just as they are for monsters who are
hit with a wand of polymorph), and would only be subject to polymorph
once they were on the ground and a second zap from a wand of polymorph
hit the space they were on. Turns out this assumption was wrong -
bypass_obj() was never in place for this kind of scenario, it was never
a condition that could happen in vanilla NetHack.

This commit makes it so if the player is zapped with a wand of
polymorph by a monster, objects on the floor in the same space the
player is standing on are affected first, so that any items the player
drops when they change form are not affected. Initially we looked at
adding the same bypass_obj() function for monster zaps player, but it
wouldn't work without a significant revamp. qt came up with a more
elegant solution. Hopefully changing the order of things in mbhit()
doesn't cause any unforseen issues.

Co-authored-by: entrez <me@entrez.cc>