Change how Infidel quest leader handles asking for the quest artifact.
Per a suggestion from Tomsod (the evil suggestion, not the 'good' one
:P) - it didn't make sense that an Infidel's quest leader would demand
that the Idol of Moloch be handed over but not the Amulet of Yendor
along with it. Or that they would demand it back and that was it -
agents of Moloch are evil. They should act that way. The change is this:
if the Infidel's quest leader feels the need to demand that the Idol be
handed over, they'll become hostile and attack you regardless of whether
you hand it over or not. If you're so weak as to volunarily hand over
the Idol, then you're not worthy of being Moloch's 'chosen one'.

If the Infidel's quest leader dies, I'm still keeping the 2nd hint with
Moloch's high priest in the Sanctum. I've changed the text so it fits
the narrative and isn't so cheesy. Plus it's a lot less work than
trying to implement the original suggestion of the quest leader
dropping a note with the 2nd hint on it.

Fixed some typo's in the quest text also.