Fix: Moloch uncursing cursed objects via prayer.
Moloch shouldn't be uncursing anything. As Moloch is the deity that
embodies all that is evil, instead of uncursing cursed objects, Moloch
will curse any blessed objects in open inventory. For the most part,
this is desireable for the Infidel role, but the player will still want
some objects to stay blessed. Not that Moloch cares.