Revamp descriptive miss messages (player vs mon, mon vs mon).
Been wanting to revamp this bit for ahile now, and have it be more
inline with how descriptive miss messages are handled with monster vs
player. Some added benefits to this revamp:

* Barding now has a proper armor class and can be treated as such. The
material it's made out of is still a factor in how much AC it provides.
* Because of the above change, a monster's barding is considered for
what armor might block an attack.
* If a monster (or player poly'd into said monster) has thick hide,
there's a descriptive miss message for that depending on what kind of
monster it is

Thanks qt for the assist with getting barding to play nice.

Co-authored-by: entrez <me@entrez.cc>