Fix: price abuse with non-human shopkeepers, player matching race.
Long-standing pricing bug that I most likely inherited from SporkHack,
not realizing exactly how much abuse could be had, milking non-human
same race shopkeepers for all of their gold. The problem was that in
get_cost() and set_cost() (buying from and selling to a shopkeeper) was
a little bit of code there that if the shopkeeper weren't human, and the
player was the same race as the shopkeeper, there was a 'professional
courtesy' multiplier that gave the player a really good deal. So good in
fact, that you could buy an object with really high charisma, lower your
charisma, then turn around a sell the same object for more than what you
paid for it. Rinse/repeat until you had all of the shopkeepers gold.
Taking that 'professional courtesy' out of the equation, the best a
player could do now with the same scheme is break even. Juice is no
longer worth the squeeze.

Thanks aosdict for noticing and mentioning it.