Make barding more like armor
Make barding act more like armor the hero can wear, with enchantment
visible after formal identification which contributes to the steed's AC.

* Barding worn by the current steed can be targeted with scrolls and
  spells which modify armor status (enchant/destroy armor, repair
  armor...).  It will evaporate when overenchanted, like normal armor
  does.
* A barded monster which triggers a rust trap may have its barding hit
  and rusted.
* Barding can be wished for with a particular enchantment, with the same
  likelihood of success as normal armor.
* A couple other minor changes made along the way (moved the weird
  barding material special bonuses from find_mac to material_bonus so
  they are included in ARM_BONUS, some other small tweaks)

I'm not very satisfied with the way 'targeted' enchant armor works with
this system: since your steed's barding isn't in your inventory, there
isn't an easy way to include it in the normal getobj prompt for
enchantment.  The current solution is to prompt specifically for barding
first if applicable, and then move on to normal armor selection if the
player declines to target the barding -- sort of similar to the old
prompt to remove a saddle/barding when #looting (which has since been
removed).  Maybe there's a better solution to this issue, not sure.