Fix: no really, fix price abuse with non-human shopkeepers, player matching race.
The first 'fix' cleaned up a few things, but players could still game
the shopkeeper by selling an object with low charisma for more than they
bought the same object with high charisma. A big part of the problem was
that get_cost() was being called inside set_cost(), which brings
charisma into play when selling an item to a shopkeeper and caused a
fair amount of weirdness. That bit has been removed, so that when you
sell an item to a shopkeeper, the price will not be affected by your
charisma. With that settled, it was still possible to game the system
(albeit not as badly), so another condition was put in get_cost() so
that an item cannot be sold to the player for less than half the base
cost (this is the same value that the player sells something to a
shopkeeper). So if a player now tries to game the system, the best
they'll do is break even. For real this time :)

In lieu of removing the blanket 'professional courtesy' bit from the
last commit, I've reintroduced same-race favoritism, but on a more
granular scale, and some races won't take your race into consideration
at all (gnomes, nymphs). I've also made sure that Izchak won't take race
into consideration at all, buying from or selling to (initially it was
just buying from).

I kinda feel like charisma *should* play a part in selling to a
shopkeeper, not just buying from them. Not going to push my luck right
now though, but I may visit this again later.