Adjust descriptive miss messages based on armor worn over other armor.
If you (or a monster) is wearing a cloak/robe over body armor, or either
of those over a t-shirt, the piece of armor that's covered up will not
be considered as a blocking piece. The added benefit of doing this, is
that this will also protect the attacker from hated material sear
damage if that piece of armor is covered up.