Priests are penalized for firing/throwing edged weapons and landing a hit.
Priests are forbidden from using edged weapons, but before this only
applied to wielded weapons. This should have applied to thrown/fired
projectiles as well. So now priests abuse their wisdom and potentially
screw up their alignment by throwing/firing the same. What's different
here vs attacking with a wielded edged weapon is that they actually need
to land a hit to anger their deity and abuse alignment. A condition was
also set so that a priest using an edged weapon ranged attack had a
chance for a misfire or slip.