Cursed wand of light causes darkness, blindingflash() fix.
Reading a cursed scroll of light produces darkness instead of light, so
should a wand of light under the same conditions. Also found a bug in
the blindingflash() routine that basically made sure it never triggered,
so that's been fixed also.

There's a bit of code formatting in weapon.c that is totally unrelated
that snuck in.