Revamp corpse revival via zombie death, zombie illness timer.
Doing this revamp for a few reasons. For starters, it brings
monster corpse revival inline with the method used in NetHack 3.7. It
also cuts down on a good bit of (now) unnecessary code, so it steamlines
things a bit. Lastly, this sets the stage for being able to add another
cool feature that was recently added to vanilla NetHack. While doing
this, I found and fixed a few minor bugs, and I also revamped how msick
is handled. It now uses a proper timeout (mdiseased could use this
treatment as well, we'll save it for another time).

Things the player will notice: monsters killed outright by a zombie
won't instantly arise as a zombie. There will be a 5-16 turn lead time
before the resurrect as one. Monsters killed from being infected by
zombie illness will behave the same way (they'll die, if they leave
a corpse and can turn into a zombie, 6-15 turn lead time).

Only issue is that racial monsters killed by a zombie will arise as a
human zombie instead of a <whatever their race was> zombie. qt had fixed
this in an older commit, but I had to break it again to get this to
where I wanted it to be. Hoping we can find a resolution to that while
keeping these recent changes intact.