Zombies can dig themselves out if buried.
This is based off of NetHack 3.7 commit c444cb3, modified. If a zombie
corpse is buried, and it's set to revive, it will dig its way out of the
ground. In vanilla NetHack, this would only happen to monsters killed by
a zombie that might revive and turn into one. Because already existing
zombies can revive if destroyed in EvilHack, regular zombie corpses get
the same treatment. I also gave the wizmode command #wizbury a hotkey
command (ctrl-b), this made it much easier to test buried zombie
corpses.

In the future I'd like to create a special buried zombie room, player
enters and all of the sudden zombies start crawling out of the ground.
Or perhaps have a chance for a grave to have a zombie corpse already
buried beneath it and have it revive (currently, zombies or mummies only
appear from a grave by the very act of digging one up). Could have some
fun with this.