New artifact (Dramborleg), minor fixes and changes to other artifacts.
Introducing a new artifact: Dramborleg. Dramborleg comes from Tolkien
lore, before the time of 'The Hobbit' or 'The Lord of the Rings'. It was
wielded by a human named Tuor, and he had used it to slay five powerful
balrogs during the Fall of Gondolin. They say no one knows which race
forged this powerful weapon, but for intents and purposes here, it's
dwarven made, and is the first dwarven artifact introduced in the game.

Dramborleg is a dwarven bearded axe made of mithril, and is lawful in
alignment. It gets a +8 to-hit and damage, and provides magic resistance
and one level of MC when wielded. A unique special capability of this
axe, is that it can destroy any balrog in one hit. Because of it's
affinity to balrogs, Dramborleg will also warn of any demons nearby.

Other changes in this commit - I noticed that the artifact Angelslayer
would never have a chance to instakill anegls and the like. The code for
that had to be moved to another section for it to work properly. The
Sceptre of Might has been changed from a heavy mace to a rod - much more
appropriate given what it is (rods use mace skill, and do slightly more
damage than a regular mace). Demonbane now has a section in the 'chance
for an instakill' bit of code - 10% chance of killing a regular demon
outright, or doing triple the damage to a demon lord/prince.