Allow various sphere-type monsters to be genocided.
Not allowing any 'e' based monsters to be genocided was something I had
pulled over from SporkHack a few years ago when I first started working
on EvilHack. No real reason to not allow at least some of them to be
genocided again. Those that are still off-limits are flaming/freezing
spheres (because of the spells that create them), magical eyes, and
beholders.