Autorecover crashed games.
From SpliceHack git commit e2bad14 (slightly modified).

I figured it's about time to allow players to be able to recover crashed
games, could be especially helpful if myself or another admin isn't
around to recover the game in a timely manor. The way it works is this -
you're playing, you encounter an error and the game crashes, sending you
back to the terminal prompt or dgl if playing on Hardfought. When you go
to start the game back up, instead of seeing 'Destroy old game?', you
first have the option to try and recover it (option 'r'). If the recover
is successful, your game is restored, typically at the very beginning of
the level you were on when the crash occurred.

Now I just need to devise a script to email me when a crash occurs.