Fix: quest leader wanting quest artifact back and it's thrown to them.
Hahah... if the quest leader was going to ask for the quest artifact
to be returned and the player threw it at them before asking, they throw
it back to the player. Nice 200 IQ way to bypass the whole handing over
the quest artifact bit :) Thanks qt for catching this one.

Co-authored-by: entrez <me@entrez.cc>