Block monster health regeneration in the Valley
Hero health regeneration is blocked in the Valley of the Dead, but the
same didn't apply to monsters (or the hero when polymorphed).  Apply the
same rule to monsters -- except for the undead, which will still recover
hit points in the Valley as before.

Since monsters regenerate HP relatively slowly, the most noticeable
aspect of this change will likely be its effect on the hero's pets
(which normally have a longer lifespan than the average monster).
Polymorphed heros will also now have to contend with the Valley regen
block (unless polymorphed into an undead monster).