Add option to filter by unidentified in item menus
Let users filter by unidentified status using 'I'.  This should work
in standard type-based inventory actions where items affected can be
filtered by type (#loot, 'D'/#droptype, etc) as well as the type-based
inventory listing of 'I'/#inventtype.

The 'I' filter uses not_fully_identified(objnam.c), so an item with a
known type may still be selected by the filter if e.g. its fooproof
status is unknown.  This could potentially leak some information about
oprops if all other info about the item is already known, so if this
turns out to be more exploitable than it seems it may need to be
revisited.

It should work with all the different menustyle options, but I only gave
a relatively cursory check to some of them, so it's definitely possible
that this will introduce some bugs or not work as expected when used
with certain combinations of menustyles and commands.