New monster - antimatter vortex.
Originally from NetHack Brass and then UnNetHack, the antimatter vortex.
This vortex has an engulfing attack just like all of the other
vorticies, but its damage type is disintegration. It also has passive
disintegration, so hitting it with objects can cause them to
disintegrate, and a bare-handed attack can kill you outright if you're
not resistant to disintegration. These can only randomly spawn in
Gehennom. They are a valid polyform, and can be genocided.

The implementation of this monster in UnNetHack caused disintegration to
kick in every single time it was struck. Probably why this monster was
eventually removed from the game. Because passive disintegration has
already been established in this variant, and it's never an always or
none kind of thing, I believe the antimatter vortex can be realized as
was intended. Striking an antimatter vortex will have a greater chance
of disintegration firing than hitting a black dragon, but the odds are
still in the attackers favor.

While engulfed, various things can happen depending on whether you're
resistant to disintegration or not, and by how much. It's basically the
same chain of events that happens if hit by a black dragons breath
attack without reflection: first your armor goes, then you.

While working on this, I noticed and fixed a bug - if the player had
partial disintegration resistance, they could still be disintegrated
from a passive attack, even if their resistance was 95%. Now, with this
commit, if your disintegration resistance is 50% or higher, you won't
ever be disintegrated outright. Eating one black dragon corpse (or tin
of the same) will get you to 50%.

We'll see how this monster goes over. Cause for concern for sure if
encountered, but by the time the player would start to see these,
chances are they already have some level of disintegration resistance.