Change minimum dexterity needed to throw Xiuhcoatl.
Mjollnir requires extremely high strength (25) to be able to throw it.
Initially to throw Xiuhcoatl, you only needed a dexterity of 18. Most
races can achieve that without any magical help, which feels not quite
right. Bump the minimum dexterity needed to toss Xiuhcoatl, so that
enchanted gauntlets of dexterity will be required (how much enchantment
is required will be dependent on your race).